+++ packages/terra-framework-docs/src/terra-dev-site/doc/slide-panel-manager/ChangeLog.3.doc.mdx
