+++ packages/terra-framework-docs/src/terra-dev-site/doc/slide-panel-manager/SlidePanelManager.1.doc.mdx
@@ -2,9 +2,9 @@ import { Badge } from 'terra-slide-panel-manager/package.json?dev-site-package';
 
 import SlidePanelManagerPropsTable from 'terra-slide-panel-manager/src/SlidePanelManager?dev-site-props-table';
 
-import SlidePanelManagerExample from '../example/SlidePanelManagerExample?dev-site-example';
-import SlidePanelManagerSquishExample from '../example/SlidePanelManagerSquishExample?dev-site-example';
-import SlidePanelManagerAccessoryExample from '../example/SlidePanelManagerAccessoryExample?dev-site-example';
+import SlidePanelManagerExample from './example/SlidePanelManagerExample?dev-site-example';
+import SlidePanelManagerSquishExample from './example/SlidePanelManagerSquishExample?dev-site-example';
+import SlidePanelManagerAccessoryExample from './example/SlidePanelManagerAccessoryExample?dev-site-example';
 
 <Badge />
