+++ packages/terra-framework-docs/src/terra-dev-site/doc/slide-panel-manager/UpgradeGuide.2.doc.mdx
