+++ packages/terra-framework-docs/src/terra-dev-site/doc/slide-panel-manager/example/SlidePanelManager.module.scss
