+++ packages/terra-framework-docs/src/terra-dev-site/test/slide-panel-manager/DisclosureComponent.jsx
