+++ packages/terra-framework-docs/src/terra-dev-site/test/slide-panel-manager/WithDisclosureContainer.test.jsx
@@ -1,19 +1,26 @@
 import React from 'react';
 import classNames from 'classnames/bind';
+import SlidePanelManager from 'terra-slide-panel-manager';
 import DisclosureComponent from './DisclosureComponent';
-import SlidePanelManager from '../../../SlidePanelManager';
 import styles from './SlidePanelManagerDefault.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const SlidePanelManagerWithDisclosureAccessory = () => (
+const DisclosureContainer = (wrappedContent) => (
+  <div data-disclosure-container className={cx('disclosure-container')}>
+    I am wrapping the content with a purple background!
+    {wrappedContent}
+  </div>
+);
+
+const SlidePanelManagerWithDisclosureContainer = () => (
   <div role="main" className={cx('content-wrapper')}>
     <SlidePanelManager
-      disclosureAccessory={<div className={cx('disclosure-accessory')}>Disclosure Accessory</div>}
+      withDisclosureContainer={DisclosureContainer}
     >
       <DisclosureComponent identifier="root-component" disclosureType="panel" renderHeaderAdapter />
     </SlidePanelManager>
   </div>
 );
 
-export default SlidePanelManagerWithDisclosureAccessory;
+export default SlidePanelManagerWithDisclosureContainer;
