--- packages/terra-slide-panel-manager/src/terra-dev-site/test/slide-panel-manager/WithDisclosureContainer.test.jsx
@@ -1,26 +1,0 @@
-import React from 'react';
-import classNames from 'classnames/bind';
-import DisclosureComponent from './DisclosureComponent';
-import SlidePanelManager from '../../../SlidePanelManager';
-import styles from './SlidePanelManagerDefault.test.module.scss';
-
-const cx = classNames.bind(styles);
-
-const DisclosureContainer = (wrappedContent) => (
-  <div data-disclosure-container className={cx('disclosure-container')}>
-    I am wrapping the content with a purple background!
-    {wrappedContent}
-  </div>
-);
-
-const SlidePanelManagerWithDisclosureContainer = () => (
-  <div role="main" className={cx('content-wrapper')}>
-    <SlidePanelManager
-      withDisclosureContainer={DisclosureContainer}
-    >
-      <DisclosureComponent identifier="root-component" disclosureType="panel" renderHeaderAdapter />
-    </SlidePanelManager>
-  </div>
-);
-
-export default SlidePanelManagerWithDisclosureContainer;
